delete win+3

+++ src/helpers/remove_folder.py
@@ -4,7 +4,7 @@ import os
 from src.helpers.validate_platform import is_win
 
 
-def remove_folder(dir_path):
+def remove_folder(dir_path, required_permission=False):
     print("Deleting ...")
     for item in os.listdir(dir_path):
         item_path = os.path.join(dir_path, item)
@@ -13,7 +13,7 @@ def remove_folder(dir_path):
             os.unlink(item_path)
         else:
             win = is_win()
-            if win:
+            if win and required_permission:
                 os.chmod(item_path, 0o700)
                 os.system(f'rmdir /s /q "{item_path}"')
             else:
